Fix the error preventing build
Building from testing currently produces the `ValueError: Newlines are not allowed` exception when `setuptools.setup()` is called, presumably due to new functionality in `setuptools`. For building on the latest setuptools, all newlines (`\n`) in the code needs to go. I have already raised the issue in the support Discord Server.